# Conflicts:
#	io.openems.edge.bridge.modbus/.settings/org.eclipse.core.resources.prefs

+++ io.openems.edge.meter.api/src/io/openems/edge/meter/api/SymmetricMeter.java
@@ -83,7 +83,7 @@ public interface SymmetricMeter extends OpenemsComponent {
 						 * Fill Min/Max Active Power channels
 						 */
 						if (value.asOptional().isPresent()) {
-							int newValue = (int)(Integer) value.get();
+							int newValue = (Integer) value.get();
 							{
 								Channel<Integer> minActivePowerChannel = channel.getComponent()
 										.channel(ChannelId.MIN_ACTIVE_POWER);
