+++ io.openems.edge.simulator/src/io/openems/edge/simulator/meter/grid/reacting/GridMeter.java
@@ -1,24 +1,5 @@
 package io.openems.edge.simulator.meter.grid.reacting;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.concurrent.CopyOnWriteArrayList;
-
-import org.osgi.service.cm.ConfigurationAdmin;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.ConfigurationPolicy;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
-import org.osgi.service.component.annotations.ReferencePolicyOption;
-import org.osgi.service.event.Event;
-import org.osgi.service.event.EventConstants;
-import org.osgi.service.event.EventHandler;
-import org.osgi.service.metatype.annotations.Designate;
-
 import io.openems.edge.common.channel.doc.Doc;
 import io.openems.edge.common.channel.doc.Unit;
 import io.openems.edge.common.component.AbstractOpenemsComponent;
@@ -29,6 +10,17 @@ import io.openems.edge.meter.api.AsymmetricMeter;
 import io.openems.edge.meter.api.MeterType;
 import io.openems.edge.meter.api.SymmetricMeter;
 import io.openems.edge.simulator.meter.MeterUtils;
+import org.osgi.service.cm.ConfigurationAdmin;
+import org.osgi.service.component.ComponentContext;
+import org.osgi.service.component.annotations.*;
+import org.osgi.service.event.Event;
+import org.osgi.service.event.EventConstants;
+import org.osgi.service.event.EventHandler;
+import org.osgi.service.metatype.annotations.Designate;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
 
 
 @Designate(ocd = Config.class, factory = true)
@@ -109,7 +101,7 @@ public class GridMeter extends AbstractOpenemsComponent
 		for (SymmetricMeter sm : this.symmetricMeters) {
 			if (sm.getMeterType() != MeterType.GRID) {
 				try {
-					powerSum += sm.getActivePower().getNextValue().get();				
+					powerSum += (Integer) sm.getActivePower().getNextValue().get();
 				} catch (NullPointerException e) {}
 			} else {
 				gridCount++;
